requirements.txt
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary # if you use Postgres
google-auth
django-cors-headers